MinCostToBuyStampAboveXDollarAmount
John wants to buy a stamp for amount of $15 but Post Office has a set of stamps and he wants to buy $15 with a combination of the stamps. He prefers not to pay too much so he wants to know the minimum amount of money he can buy with the given set of stamps, but has to be above $15.